# apps/docs/public/openapi.yaml
openapi: 3.0.0
info:
  title: MotoReco API Documentation
  version: 0.0.1

servers:
  - url: http://localhost:3002
    description: ローカル開発環境

tags:
  - name: user
    description: ユーザー情報関連のAPI
  - name: bikes
    description: バイク情報関連のAPI
  - name: user-bikes
    description: ユーザー所有バイク関連のAPI

paths:
  /api/v1/user/profile:
    get:
      summary: ユーザプロフィールの取得
      tags:
        - user
      responses:
        '200':
          description: プロフィール取得成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserProfile'
                  message:
                    type: string
                    example: プロフィール取得成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      summary: ユーザプロフィールの更新
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateProfile'
      responses:
        '200':
          description: プロフィール更新成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserProfile'
                  message:
                    type: string
                    example: プロフィール更新成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/user/auth/register:
    post:
      summary: ユーザー登録（新規登録または既存ユーザー取得）
      description: Firebase認証後の初回登録エンドポイント。既にMAuthProviderにレコードが存在する場合は既存ユーザー情報を返却します。
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: ユーザー登録成功（新規作成または既存ユーザー）
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserProfile'
                  message:
                    type: string
                    example: ユーザー登録成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/bikes/manufacturers:
    get:
      summary: バイクメーカー一覧の取得
      tags:
        - bikes
      responses:
        '200':
          description: メーカー一覧取得成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      manufacturers:
                        type: array
                        items:
                          $ref: '#/components/schemas/ManufacturerDetail'
                  message:
                    type: string
                    example: メーカー一覧取得成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/bikes/search:
    get:
      summary: 登録可能なバイクの検索
      tags:
        - bikes
      parameters:
        - in: query
          name: mf-op
          schema:
            type: string
            enum:
              - eq
              - ne
              - in
        - in: query
          name: mf
          schema:
            type: string
            description: バイクメーカーID（複数指定可能、カンマ区切り、 `mf-op`で指定したオペレーターに従う）
        - in: query
          name: model-name
          schema:
            type: string
            description: モデル名（部分一致検索）
        - in: query
          name: displacement-min
          schema:
            type: integer
            description: 排気量の最小値（cc）
        - in: query
          name: displacement-max
          schema:
            type: integer
            description: 排気量の最大値（cc）
        - in: query
          name: model-year-min
          schema:
            type: integer
            description: モデル公開年の最小値
        - in: query
          name: model-year-max
          schema:
            type: integer
            description: モデル公開年の最大値
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            description: ページ番号（1始まり）
        - in: query
          name: per-size
          schema:
            type: integer
            default: 20
            maximum: 100
            description: 1ページあたりの件数
        - in: query
          name: sort-by
          schema:
            type: string
            enum:
              - model-year
              - displacement
              - modelYear
            description: ソート対象のフィールド
        - in: query
          name: sort-order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
            description: ソート順序
      responses:
        '200':
          description: バイク検索成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      bikes:
                        type: array
                        items:
                          $ref: '#/components/schemas/BikeDetail'
                  message:
                    type: string
                    example: バイク検索成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/user-bike/register:
    post:
      summary: マイユーザーバイクの登録
      tags:
        - user-bikes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBikeRegisterDetail'
      responses:
        '201':
          description: ユーザー所有バイク登録成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserBikeDetail'
                  message:
                    type: string
                    example: ユーザー所有バイク登録成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/user-bike/bikes:
    get:
      summary: ユーザー所有バイクの一覧取得
      tags:
        - user-bikes
      parameters:
        - in: query
          name: sort-by
          schema:
            type: string
            enum:
              - created-at
              - updated-at
            default: updated-at
            description: ソート対象のフィールド
        - in: query
          name: sort-order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
            description: ソート順序
      responses:
        '200':
          description: ユーザー所有バイク一覧取得成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      bikes:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserBikeDetail'
                  message:
                    type: string
                    example: ユーザー所有バイク一覧取得成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/user-bike/bike/{my-user-bike-id}:
    get:
      summary: ユーザー所有バイクの詳細取得
      tags:
        - user-bikes
      parameters:
        - in: path
          name: my-user-bike-id
          required: true
          schema:
            type: string
          description: ユーザー所有バイクID
      responses:
        '200':
          description: ユーザー所有バイク詳細取得成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserBikeDetail'
                  message:
                    type: string
                    example: ユーザー所有バイク詳細取得成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    patch:
      summary: ユーザー所有バイクの情報更新
      tags:
        - user-bikes
      parameters:
        - in: path
          name: my-user-bike-id
          required: true
          schema:
            type: string
          description: ユーザー所有バイクID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBikeUpdateDetail'
      responses:
        '200':
          description: ユーザー所有バイク情報更新成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserBikeDetail'
                  message:
                    type: string
                    example: ユーザー所有バイク情報更新成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/user-bike/bike/{my-user-bike-id}/fuel-logs:
    get:
      summary: ユーザー所有バイクの燃料ログ一覧取得
      tags:
        - user-bikes
      parameters:
        - in: path
          name: my-user-bike-id
          required: true
          schema:
            type: string
          description: ユーザー所有バイクID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            description: ページ番号（1始まり）
        - in: query
          name: per-size
          schema:
            type: integer
            default: 20
            maximum: 100
            description: 1ページあたりの件数
        - in: query
          name: sort-by
          schema:
            type: string
            enum:
              - refueled-at
              - mileage
        - in: query
          name: sort-order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
            description: ソート順序
      responses:
        '200':
          description: 燃料ログ一覧取得成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MyUserBikeFuelLogDetail'
                  message:
                    type: string
                    example: 燃料ログ一覧取得成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      summary: ユーザー所有バイクの燃料ログ登録
      tags:
        - user-bikes
      parameters:
        - in: path
          name: my-user-bike-id
          required: true
          schema:
            type: string
          description: ユーザー所有バイクID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MyUserBikeFuelLogRegisterDetail'
      responses:
        '201':
          description: 燃料ログ登録成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/MyUserBikeFuelLogDetail'
                  message:
                    type: string
                    example: 燃料ログ登録成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    patch:
      summary: ユーザー所有バイクの燃料ログ更新
      tags:
        - user-bikes
      parameters:
        - in: path
          name: my-user-bike-id
          required: true
          schema:
            type: string
          description: ユーザー所有バイクID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MyUserBikeFuelLogUpdateDetail'
      responses:
        '200':
          description: 燃料ログ更新成功
          headers:
            Authorization:
              $ref: '#/components/headers/Authorization'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/MyUserBikeFuelLogDetail'
                  message:
                    type: string
                    example: 燃料ログ更新成功
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'

  /api/v1/health:
    get:
      summary: ヘルスチェック
      responses:
        '200':
          description: サービス稼働中
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                  data:
                    type: object
                    example: null

components:
  headers:
    Authorization:
      description: Bearerトークンを指定します。
      example: Bearer your_token_here
      required: true
      schema:
        type: string

  schemas:
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 50
    UserUpdateProfile:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
    UserRegister:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: ユーザー名（1-50文字）
    ManufacturerDetail:
      type: object
      properties:
        manufacturerId:
          type: string
        name:
          type: string
        nameEn:
          type: string
        country:
          type: string
    BikeDetail:
      type: object
      properties:
        bikeId:
          type: string
        manufacturerId:
          type: string
        manufacturer:
          type: string
        modelName:
          type: string
        displacement:
          type: number
          format: float
          description: 排気量（cc）
        modelYear:
          type: integer
          description: モデル公開年
    UserBikeDetail:
      type: object
      properties:
        userBikeId:
          type: string
        myUserBikeId:
          type: string
        manufacturerName:
          type: string
          nullable: true
        bikeId:
          type: string
          nullable: true
        modelName:
          type: string
          nullable: true
        nickname:
          type: string
          nullable: true
          default: null
          maxLength: 50
        purchaseDate:
          type: string
          description: 購入日（ISO 8601形式）
          nullable: true
        purchasePrice:
          type: integer
          description: 購入価格（円）
          nullable: true
        purchaseMileage:
          type: integer
          description: 購入時走行距離（km）
          nullable: true
        totalMileage:
          type: integer
          description: 総走行距離（km）
          nullable: true
          default: 0
        displacement:
          type: number
          format: float
          description: 排気量（cc）
        modelYear:
          type: integer
          description: モデル公開年
          nullable: true
          example: 2025
        createdAt:
          type: string
          description: 登録日時（ISO 8601形式）
        updatedAt:
          type: string
          description: 最終更新日時（ISO 8601形式）
    UserBikeRegisterDetail:
      type: object
      anyOf:
        - required:
            - bikeId
        - required:
            - displacement
      properties:
        bikeId:
          type: string
          nullable: true
        displacement:
          type: number
          format: float
          description: 排気量（cc）。bikeId未指定の場合は必須。
        serialNumber:
          type: string
          nullable: true
          maxLength: 100
        nickname:
          type: string
          nullable: true
          default: null
          maxLength: 50
        purchaseDate:
          type: string
          format: date-time
          description: 購入日（ISO 8601形式）
          nullable: true
        purchasePrice:
          type: integer
          description: 購入価格（円）
          nullable: true
        purchaseMileage:
          type: integer
          description: 購入時走行距離（km）
          nullable: true
        totalMileage:
          type: integer
          description: 総走行距離（km）
          nullable: true
          default: 0
    UserBikeUpdateDetail:
      type: object
      properties:
        nickname:
          type: string
          nullable: true
          default: null
          maxLength: 50
        purchaseDate:
          type: string
          description: 購入日（ISO 8601形式）
          nullable: true
        purchaseMileage:
          type: integer
          description: 購入時走行距離（km）
          nullable: true
        purchasePrice:
          type: integer
          description: 購入価格（円）
          nullable: true
        totalMileage:
          type: integer
          description: 総走行距離（km）
          nullable: true
          default: 0
    MyUserBikeFuelLogDetail:
      type: object
      properties:
        fuelLogId:
          type: string
        refueledAt:
          type: string
          description: 給油日時（ISO 8601形式）
        mileage:
          type: integer
          description: 給油時走行距離（km）
        amount:
          type: number
          format: float
          description: 給油量（リットル）
        totalPrice:
          type: number
          format: integer
          description: 合計価格（円）
    MyUserBikeFuelLogRegisterDetail:
      type: object
      required:
        - refueledAt
        - mileage
        - amount
        - totalPrice
      properties:
        refueledAt:
          type: string
          description: 給油日時（ISO 8601形式）
        mileage:
          type: integer
          description: 給油時走行距離（km）
        amount:
          type: number
          format: float
          description: 給油量（リットル）
        totalPrice:
          type: number
          format: integer
          description: 合計価格（円）
        updateTotalMileage:
          type: boolean
          description: 燃料ログ登録後にユーザー所有バイクの総走行距離を更新するかどうか。更新する値が現在の総走行距離よりも小さい場合は更新されません。
          default: false
    MyUserBikeFuelLogUpdateDetail:
      type: object
      required:
        - fuelLogId
      properties:
        fuelLogId:
          type: string
          description: 燃料ログID
        refueledAt:
          type: string
          description: 給油日時（ISO 8601形式）
        mileage:
          type: integer
          description: 給油時走行距離（km）
        amount:
          type: number
          format: float
          description: 給油量（リットル）
        totalPrice:
          type: number
          format: integer
          description: 合計価格（円）

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        errorCode:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        detail:
          type: object
          nullable: true
          description: 詳細情報（任意）

    ErrorCode:
      type: string
      enum:
        - INVALID_REQUEST
        - VALIDATION_ERROR
        - AUTH_FAILED
        - USER_NOT_REGISTERED
        - NOT_FOUND
        - SERVER_ERROR

    MaintenanceType:
      type: string
      enum:
        # ブレーキ装置
        - BRAKE_FLUID
        - FRONT_BRAKE_PAD
        - REAR_BRAKE_PAD
        - MASTER_CYLINDER_CUP
        - BRAKE_CALIPER_SEAL
        - BRAKE_CABLE

  responses:
    Error400:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: 認証失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: ユーザー未登録
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error500:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'